1.0.0
- rename gem to `tablerize`
- bump version number
- add RSpec
