+++ .gitignore
@@ -1,1 +1,1 @@
-/yaml-tablerize-*.gem
+/tablerize-*.gem
