+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2013 IFTTT Inc
+Copyright (c) 2015 IFTTT Inc
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
