+++ examples/example-1.html
@@ -1,0 +1,9 @@
+<table class="statistics-table nsa-surveillance-details"><tbody>
+<tr class="table-header"><td class="statistics-table-authority">Legal Authority</td><td class="statistics-table-num_orders">Annual Number of Orders</td><td class="statistics-table-num_targets">Estimated Number of Targets Affected</td></tr>
+<tr><td class="statistics-table-authority"><strong>FISA Orders</strong><br />
+Based on probable cause (Title I and III of FISA, Sections 703 and 704 of FISA)</td><td class="statistics-table-num_orders">1,167 orders</td><td class="statistics-table-num_targets">1,144</td></tr>
+<tr><td class="statistics-table-authority"><strong>Section 702</strong><br />
+of FISA</td><td class="statistics-table-num_orders">1 order</td><td class="statistics-table-num_targets">89,138</td></tr>
+<tr><td class="statistics-table-authority"><strong>FISA Pen Register/Trap and Trace</strong><br />
+(Title IV of FISA)</td><td class="statistics-table-num_orders">131 orders</td><td class="statistics-table-num_targets">319</td></tr>
+</tbody></table>
