+++ examples/example-2.html
@@ -1,0 +1,13 @@
+<table class="http-spec-exchange another-class"><tbody>
+<tr><td class="http-key http-spec-exchange-k">Method</td><td class="http-value http-spec-exchange-v">GET</td></tr>
+<tr><td class="http-key http-spec-exchange-k">Parameters</td><td class="http-value http-spec-exchange-v">
+<table class="http-spec-params"><tbody>
+<tr><td class="http-spec-params-k"><code>client_id</code></td><td class="http-spec-params-v"><p>A client ID for your service as set in your configuration.</p>
+
+<p>a new line, wow! Let’s see regular Markdown tables do that…</p></td></tr>
+<tr><td class="http-spec-params-k"><code>type</code></td><td class="http-spec-params-v"><code>code</code></td></tr>
+<tr><td class="http-spec-params-k"><code>state</code></td><td class="http-spec-params-v">An anti-forgery token provided by the API.</td></tr>
+<tr><td class="http-spec-params-k">redirect_uri</td><td class="http-spec-params-v"><code>https://example.com/api/{{your_service}}/authorize</code></td></tr>
+</tbody></table>
+</td></tr>
+</tbody></table>
