+++ lib/tablerize/version.rb
@@ -1,0 +1,3 @@
+module Tablerize
+  VERSION = "1.0.0"
+end
