--- lib/yaml-tablerize/version.rb
@@ -1,3 +1,0 @@
-module YamlTablerize
-  YAML_TABLERIZE_VERSION = "0.2.1"
-end
