+++ spec/spec_helper.rb
@@ -1,0 +1,1 @@
+require_relative '../lib/tablerize'
